% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VDJ.utils.R
\name{Ab1toAIRR}
\alias{Ab1toAIRR}
\title{unzip Eurofins-based results and run all steps from QC to final AIRR formatted datatable}
\usage{
Ab1toAIRR(
  files,
  primers = c("IgG", "IgM", "IgL", "IgK", "Mix"),
  trim_cutoff = NULL,
  QC = TRUE,
  igblast = TRUE,
  seq_type = c("Ig", "TCR"),
  organism = c("human", "mouse", "rabbit", "rat", "rhesus_monkey"),
  igblast_log = TRUE,
  igblast_dir = "~/share/igblast/",
  imgt_dir = "~/share/germlines/imgt/",
  igblast_method = c("AssignGenes", "runIgblastn"),
  update_c_call = TRUE,
  SHM = TRUE,
  full_seq_aa = TRUE,
  update_info = FALSE,
  verbose = TRUE,
  return_db = FALSE,
  save = c("png", "html"),
  reticulate_py_env = "~/Library/r-miniconda-arm64/envs/r-reticulate/bin/python",
  nproc = NULL
)
}
\arguments{
\item{files}{a vector of full file paths to Ab1.zip files}

\item{primers}{primers used (one of IgG, IgM, IgL, IgK or Mix)}

\item{trim_cutoff}{cutoff thresholds to use [if set to NULL, default to automatic cutoffs for each primers: "IgG" = 300, "IgM" = 275, "IgK" = 400, "IgL"= 275, "Mix" = 275, "Undefined" = 275]}

\item{QC}{whether to run initial QC part}

\item{igblast}{whether to run igblast}

\item{seq_type}{one of Ig or TCR (passed to igblast)}

\item{organism}{one of human, mouse,}

\item{igblast_dir}{location of the igblast database}

\item{igblast_method}{whether to run runAssignGenes or runIgblastn}

\item{update_c_call}{whether to update c_call using runBlastnC}

\item{SHM}{whether to calculate mutations in V gene}

\item{full_seq_aa}{whether to reconstruct full AA sequences for each sequence}

\item{update_info}{whether to only import new info from updated template.info.xlsx file(s); prior steps (QC and igblast) should have been runned before.}

\item{verbose}{whether to write to the console}

\item{return_db}{whether to return the final productive sequences data table}

\item{save}{passed to runAb1QC, whether to save png or html plots}

\item{reticulate_py_env}{if you want to force a precise python environment to use for plotly. [default: NULL]}

\item{nproc}{number of processors to use, if NULL, will be automatically set based on available processors.}

\item{reference_dir}{location of the imgt gap database}
}
\value{
A data table containing all sequences passing QC and igblast.
All QC analysis plots, intermediate tsv files with sequences passing and failing igblast and an excel worksheet containing sheets with AIRR formatted data table for sequences failing initial QC, failing igblast, non-productive and passing all three steps.
}
\description{
\code{Ab1toAIRR}
}
