% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VDJ.utils.R
\name{addAIRRmetadata}
\alias{addAIRRmetadata}
\title{Adds AIRR metadata of a seurat object}
\usage{
addAIRRmetadata(
  sc,
  vdj_db = NULL,
  seq_type = c("Ig", "TCR"),
  import = c("bcr_info", "tcr_info", "clone_id", "expanded_clone", "sequence_id",
    "consensus_count", "umi_count", "locus", "productive", "v_call", "d_call", "j_call",
    "c_call", "junction", "junction_aa", "junction_length", "mu_count", "mu_count_cdr_r",
    "mu_count_cdr_s", "mu_count_fwr_r", "mu_count_fwr_s", "is.VDJ_doublet",
    "is.VDJ_doublet.confidence", "is.nonB_VDJ_doublet", "is.nonB_VDJ_doublet.confidence"),
  locus = "locus",
  cell_id = "cell_id",
  clone_id = "clone_id",
  split.by = NULL,
  commun_columns = c("bcr_info", "tcr_info", "clone_id", "expanded_clone",
    "is.VDJ_doublet", "is.VDJ_doublet.confidence", "is.nonB_VDJ_doublet",
    "is.nonB_VDJ_doublet.confidence"),
  chains = list(TCR = list(tcra = "TCRA", tcrb = "TCRB", tcrg = "TCRG", tcrd = "TCRD"),
    BCR = list(igheavy = "IGH", iglight = c("IGL", "IGK")))
)
}
\arguments{
\item{sc}{a seurat object on which to input vdj data, should include a cell_id column matching the cell_id column in the vdj_db dataframe.}

\item{vdj_db}{an AIRR formatted dataframe containing bcr (heavy and light chains) or tcr (TCRA, TCRB, TCRG or TCRD) sequences. Should contain only one chain for each type per cell_id, if not run resolveMultiHC() first.}

\item{import}{list of columns to import, if set to "all" will import all columns not already included in the seurat object. If "clone_id" is present will calculate clone size and frequencies.}

\item{locus}{name of column containing locus values.}

\item{cell_id}{name of the column containing cell identifier.}

\item{clone_id}{name of the column containing cell identifier.}

\item{split.by}{name of column to use to group sequence when calculating clone size and frequencies.}

\item{commun_columns}{name of columns commun between different chains. Will not be rename when all other chain specific columns will be renamed "chain_old name"}

\item{chains}{description of chain pairing. list names will be used for renaming.}

\item{type}{either BCR or TCR, by default BCR.}
}
\value{
a seurat object with additional metadata imported from the vdj_db dataframe.
}
\description{
\code{AddAIRRmetadata} adds the immune receptor information to the metadata of a seurat object
}
\details{
The provided df is expected to follow the AIRR format for column nomenclatures, and to contain a "clone_id" column.
Cell_ids with duplicated heavy and light chain should also have been excluded during prior steps of the pipeline. By default will be excluded here.
Before using AddAIRRmetadata() ensure the cell_ids of the seurat object match the "cell_id" in the AIRR formatted dataframe.
By default, if "clone_id" column exists for tcra and tcrb or for heavy_chain this function also calculates the size and frequencies of the clonotypes (as defined in the clone_id column).
(use the "groupBy =" argument, by default NULL will return frequency among all cells in the dataset, if set to NULL, no frequency is returned).
Grouping in done at the sc level post incorporation of the VDJ info so all columns from both sc and import can be used.
}
