% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VDJ.plots.R
\name{DonutPlotClonotypes3D}
\alias{DonutPlotClonotypes3D}
\title{Plots multiple donut plots and save them as pdf}
\usage{
DonutPlotClonotypes3D(
  db,
  split.by = NULL,
  prefix = NULL,
  use_chain = "IGH",
  locus = "locus",
  cell_id = "cell_id",
  clone_id = "clone_id",
  groups_to_plot = "all",
  col.line = "black",
  plots_folder = "Donut_plots",
  highlight = c("shared", "clone_size", "clone_rank"),
  highlight_col = list(shared = c("Paired", "random"), clone_size = c("Set1"), clone_rank
    = c("Set1")),
  external_bar = c("expanded", "top5", "none"),
  productive = "productive",
  productive_only = TRUE
)
}
\arguments{
\item{db}{an AIRR formatted dataframe containing bcr (heavy and light chains) or tcr (TCRA, TCRB, TCRG or TCRD) sequences. Should contain only one chain for each type per cell_id, if not run resolveMultiHC() first.}

\item{split.by}{name of column to use to group sequence when calculating clone size and frequencies.}

\item{prefix}{prefix to use for saved files}

\item{use_chain}{which chain to use [default: "IGH"], each cell should only have one contig for this chain}

\item{locus}{name of column containing locus values.}

\item{cell_id}{name of the column containing cell identifier.}

\item{clone_id}{name of the column containing cell identifier.}

\item{groups_to_plot}{which groups to plot}

\item{col.line}{color for lines in donut plot}

\item{plots_folder}{name for export folder [default: "Donut_plots"]}

\item{highlight}{whether to highlight shared clones or clones based on clone_size or clone_rank. Also possible to pre-classify cells prior and set that column in highlight + a color scheme in highlight_col}

\item{highlight_col}{color scheme to use for highlighed clones}

\item{external_bar}{whether to add an external bar for "expanded" or "top5" clones.}

\item{productive}{name of column containing productive calls.}

\item{productive_only}{whether to exclude non productive sequences [default: TRUE]}
}
\value{
multiple pdf donut plots
}
\description{
\code{DonutPlotClonotypes3D} Plots multiple donut plots and save them as pdf
}
\details{
db should be an AIRR formated database with a defined "clone_id" column
Defines clones to plot and colors based on clones properties:
Expanded clone = clone size >1 in one donor
Shared clone = clone found in at least two donors
Persisting clone = clone found in at least two time points and/or pop
Color choice:
Single Greys scale based on clone size in that donor (to account for sampling): "white" = singlets, "black"= biggest clone in all donors
Color for shared clones between donors
highlight = c("expanded", "top5")
Consistent color throughout time-points for persisting-expanded clones
Grey for non-persisting expanded clones
White for non expanded clones and regroup all sequences found only once in the dataset (non-persisting/non-expanded)
}
