% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VDJ.utils.R
\name{flagBdoublets}
\alias{flagBdoublets}
\title{check for cases where two IGH or two IGH and two light chains are found in a cell and could represent a B cell doublet.}
\usage{
flagBdoublets(
  db,
  analysis_name = "All_sequences",
  use_chain = c("all", "IGH"),
  split.by = NULL,
  variable_cutoff = TRUE,
  low_cutoff = 10,
  high_cutoff = 250,
  output = TRUE,
  output_folder = "VDJ_QC",
  cell_id = "cell_id",
  locus = "locus",
  umi_count = "umi_count"
)
}
\arguments{
\item{db}{a AIRR formatted dataframe containing heavy and light chain sequences, cell_ids and umi_counts.}

\item{analysis_name}{name to use for outputs prefixes.}

\item{use_chain}{which chain to use to flag VDJ doublets, can de set to "IGH" or "all" [default = "all"].}

\item{split.by}{name of the column in the dataframe to use to split the dataset prior to calculating variable cut-offs}

\item{variable_cutoff}{whether to use variable cutoffs [default = TRUE, cutoffs are automatically calculated based on summary counts in the database to account for potential sequencing bias (1st quartile and max of 10 or 1/10 of median).].}

\item{low_cutoff}{cut_off for low probability heavy chain doublets (only used if variable_cutoff = FALSE).}

\item{high_cutoff}{cut_off for high probability heavy chain doublets (only used if variable_cutoff = FALSE).}

\item{output}{whether to output graphs with umi_counts for dominant versus second IGH VDJ contig and the recap excel workbook. Can be set to none, partial (only pdf + log file with QC parameters) or total (pdf + log file with QC parameters + tsv file with sequences).}

\item{output_folder}{name of the folder in which graph and recap excel workbooks will be saved [default = "VDJ_QC"].}

\item{cell_id}{name of the column containing cell identifier.}

\item{locus}{name of column containing locus values.}

\item{umi_count}{name of the column containing the number of unique molecules (UMI) for this contig. Previously called "duplicate_count" in an earlier AIRR standard}
}
\value{
the original data frame with for the following additional columns: "is.VDJ_doublet" and "is.IGH_doublet.confidence" (use_chain = "IGH") or "is.IGH_doublet.confidence", "is.IGK_doublet.confidence" and "is.IGL_doublet.confidence" (use_chain = "all")
as well as graphs plotting the primary versus "second_" umi_counts by locus for all cell_ids colored by "is."locus"_VDJ_doublet.confidence".
}
\description{
\code{flagBdoublets} flag potential B cell doublets.
}
\details{
first group by locus, split.by and calculate variable cutoffs (upper = max(10, 1st quartile of distribution); lower = max(5, median(distribution)/10))
then calculate doublet.confidences for all locus: a high confidence doublet is defined as umi_count > 10 and !(umi_count>3*second_umi_count & second_umi_count < upper_cutoff), for low_confidence doublets, we move to second_umi_count < lower_cutoff.
then is.VDJ_doublet is defined as that cell must be high for at least one of the chain (and low for the others if multiple chains are used).
dependencies: dplyr, tidyr, ggplot2
}
