% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VDJ.utils.R
\name{flagNonBdoublets}
\alias{flagNonBdoublets}
\title{Predicts nonB-B doublets}
\usage{
flagNonBdoublets(
  db,
  scRNAseq.tech = c("10X", "BD"),
  split.by = NULL,
  output = TRUE,
  analysis_name = "All_sequences",
  output_folder = "VDJ_QC/",
  variable_cutoff = TRUE,
  low_cutoff = 10,
  high_cutoff = 250,
  azimuth.ref = c("pbmcref", "tonsilref", "bonemarrowref"),
  azimuth.column = list(pbmcref = "predicted.celltype.l2", tonsilref =
    "predicted.cell_type.l1", bonemarrowref = "predicted.celltype.l2"),
  azimuth.Bcelltypes = list(pbmcref = c("B intermediate", "B memory", "B naive",
    "Plasmablast"), tonsilref = c("B activated", "B memory", "B naive", "preGCB", "PB",
    "PC", "PC/doublet", "preMBC/doublet", "prePB", "Cycling DZ GCB", "DZ GCB",
    "DZtoLZ GCB transition", "FCRL4/5+ B memory", "LZ GCB", "LZtoDZ GCB transition"),
    bonemarrowref = c("Memory B", "Naive B", "Plasma", "pro B", "pre B",
    "transitional B")),
  clone_id = "clone_id",
  cell_id = "cell_id",
  locus = "locus",
  heavy = "IGH",
  assay = "assay",
  productive = "productive",
  complete_vdj = "complete_vdj",
  sequence_id = "sequence_id",
  umi_count = "umi_count",
  consensus_count = "consensus_count",
  junction = "junction",
  junction_aa = "junction_aa",
  sequence = "sequence",
  v_call = "v_call",
  d_call = "d_call",
  j_call = "j_call",
  c_call = "c_call"
)
}
\arguments{
\item{db}{an AIRR formatted data frame containing heavy and light chain sequences, cell_ids and umi_counts.}

\item{scRNAseq.tech}{vector with names of scRNA-seq technologies used in the assay column, i.e. those from which we have .}

\item{split.by}{name of the column in the seurat object to use to split the dataset. Used both in cases it is needed to filter IGH contigs and for variable cutoffs definitions and plotting.}

\item{output}{whether to output graphs with umi_counts.}

\item{output_folder}{name of the folder in which graph and recap excel workbooks will be saved [default = "VDJ_QC"].}

\item{variable_cutoff}{whether cutoffs are automatically calculated based on summary counts in the database to account for potential sequencing bias (1st quartile and max of 10 or 1/10 of median).
else, cutoffs should be provided or preset cutoffs will be used.}

\item{low_cutoff}{cut_off for low probability heavy chain doublets.}

\item{high_cutoff}{cut_off for high probability heavy chain doublets.}

\item{azimuth.ref}{which azimuth ref was used for mapping of scRNA-seq dataset: one of "pbmcref", "tonsilref", "bonemarrowref" [default datasets as of Feb 2025] [TODO: add gut dataset?]}

\item{azimuth.column}{which output column from azimuth should be used [default values as of Feb 2025]}

\item{azimuth.Bcelltypes}{which clusters in the azimuth output correspond to B/PC cells populations likely to carry a BCR [default values as of Feb 2025].}

\item{clone_id}{name of the column containing clone identifier [only use if need to filter light chain, if absent defaulting to cell_id].}

\item{cell_id}{name of the column containing cell identifier.}

\item{locus}{name of column containing locus values.}

\item{heavy}{value of heavy chains in locus column. All other values will be treated as light chains.}

\item{productive}{name of the column containing the info whether a given sequence is productive.}

\item{sequence_id}{name of the column containing sequence identifier.}

\item{umi_count}{name of the column containing the number of unique molecules (UMI) for this contig. Previously called "duplicate_count" in an earlier AIRR standard}

\item{consensus_count}{name of the column containing the number of reads for this contig (usually called consensus_count)}

\item{junction}{name of the column containing identified junction in nucleotide format.}

\item{junction_aa}{name of the column containing identified junction in amino-acid format.}

\item{sequence}{name of the column containing the original sequence.}

\item{v_call}{name of the column containing V-segment allele assignments. All
entries in this column should be identical to the gene level.}

\item{d_call}{name of the column containing D-segment allele assignments. All
entries in this column should be identical to the gene level.}

\item{j_call}{name of the column containing J-segment allele assignments. All
entries in this column should be identical to the gene level.}

\item{c_call}{name of the column containing Constant region assignments. All
entries in this column should be identical to the gene level.}
}
\value{
the original data frame with for two new columns: "is.nonB_VDJ_doublet" and "is.nonB_VDJ_doublet.confidence" as well as graphs plotting the IGH and IGL/K umi_counts for all cell_ids colored by "is.nonB_VDJ_doublet.confidence".
}
\description{
\code{flagNonBdoublets} Predicts nonB-B doublets based on azimuth predictions and IGH-IGL/K contigs umi_counts.
}
\details{
If multiple IGH or IGL/K contigs, will first run resolveMultuiHC()/resolveMultiLC().
Then will define variable cutoffs based on librairie(s) distribution, and assign a probability of being a true nonB-B doublet for each cell not predicted as being a B cell by azimuth.
High probably nonB-B doublets are define as having both heavy and light chain in the top three quartiles of IGH/IGL/IGK expression in the dataset yet not being mapped to a B cell cluster by azimuth.

https://azimuth.hubmapconsortium.org/
for "bonemarrowref", use predicted.celltype.l2 and ! %in% c("Memory B", "Naive B", "Plasma", "pro B, "pre B", "transitional B")
for "pbmcref", use predicted.celltype.l2 and ! %in% c("B intermediate", "B memory", "B naive", "Plasmablast")
for "tonsilref v2", use predicted.celltype.l1 and ! %in% c("B activated", "B memory", "B naive", "preGCB", "PB", "PC", "PC/doublet", "preMBC/doublet", "prePB", "Cycling DZ GCB","DZ GCB","DZtoLZ GCB transition","FCRL4/5+ B memory","LZ GCB","LZtoDZ GCB transition")

dependencies: dplyr, tidyr, ggplot2 and openxlsx
}
