% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VDJ.utils.R
\name{flagVDJdoublets}
\alias{flagVDJdoublets}
\title{calls successively flagBdoublets and flagNonBdoublets}
\usage{
flagVDJdoublets(
  db,
  analysis_name = "All_sequences",
  homotypic = TRUE,
  heterotypic = TRUE,
  seq_type = c("Ig", "TCR"),
  heavy = NULL,
  light = NULL,
  output = TRUE,
  output_folder = "VDJ_QC",
  split.by = NULL,
  ref = NULL,
  ref.column = NULL,
  ref.Bcelltypes = NULL,
  ref.Tcelltypes = NULL,
  save_plot = c("pdf", "png"),
  save_tsv = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{db}{a AIRR formatted dataframe containing heavy and light chain sequences, cell_ids and umi_counts.}

\item{analysis_name}{name to use for outputs prefixes.}

\item{homotypic}{whether to run homotypicVDJdoublets()}

\item{heterotypic}{whether to run heterotypicVDJdoublets()}

\item{seq_type}{type of VDJ sequence ("Ig" or "TCR" to match igblastb requirements)}

\item{heavy}{heavy chains to be used. if set to NULL, will default to IGH for Ig and TRB/TRD for TCR.}

\item{light}{light chains to be used. if set to NULL, will default to IGL/IGK for Ig and TRA/TRG for TCR.}

\item{output}{whether to output graphs with umi_counts for dominant versus second IGH VDJ contig and the recap excel workbook. Can be set to none, partial (only pdf + log file with QC parameters) or total (pdf + log file with QC parameters + tsv file with sequences).}

\item{output_folder}{name of the folder in which graph and recap excel workbooks will be saved [default = "VDJ_QC"].}

\item{split.by}{which column to use to group cells when learning dataset-specific distributions}

\item{ref}{which ref was used for mapping of scRNA-seq dataset: one of "azimuth.pbmcref", "azimuth.tonsilref", "azimuth.bonemarrowref" [default azimuth datasets as of June 2025] or "internal", if using your own clustering information [TODO: add gut dataset?]; 
to add you own clustering info, set ref = "internal", provide the clustering column name in ref.column and corresponding B or T cell clusters in ref.Bcelltypes = c("C1", "C3"...) or ref.Tcelltypes = c("C4", "C5"...)...}

\item{ref.column}{which output column to use for cluster information [default azimuth values provided as of June 2025]}

\item{ref.Bcelltypes}{which clusters in the azimuth output correspond to B/PC cells populations likely to carry a BCR [default values provided as of June 2025].}

\item{ref.Tcelltypes}{which clusters in the azimuth output correspond to T cells populations likely to carry a TCR [default values provided as of June 2025].}

\item{save_plot}{format to save the final ggplot2 object. "pdf" or "png". if set to any other value, no plot will be saved.}

\item{verbose}{whether to print recap informations to the console}

\item{...}{arguments to pass to homotypicVDJdoublets and heterotypicVDJdoublets.}
}
\description{
\code{flagVDJdoublets} flag potential B and nonB doublets.
}
