% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowJo.utils.R
\name{FlowJoBarcode}
\alias{FlowJoBarcode}
\title{creates FlowJO compatible barcodes and back}
\usage{
FlowJoBarcode(db, input = "10X", orig.idents = NULL, end = "-1")
}
\arguments{
\item{db}{an AIRR formatted data frame containing heavy and light chain sequences to which the imported FlowJo export data will be appended. if NULL, returning only the merged FlowJo export data}

\item{input}{10X or FlowJo (TODO add compatibility to BD Rhapsody)}

\item{orig.idents}{column in FJ_files to use to find the file directory}

\item{end}{optional value to add or remove after the actual barcode [default for 10X: -1]}
}
\description{
\code{FlowJoBarcode} creates FlowJO compatible barcodes and back
}
\details{
useful to import 10X ADT counts (or other) into FlowJo for analysis and fine gating of cells of interest for future import into FlowJo.
input can be "10X" or "FlowJo"
10X: will replace the "cell_id" column by 1 orig.idents (if !NULL) and 4 numbered id columns
FlowJo: will recreate the original 10X barcode based on the orig.idents (if !NULL) and numbered id columns
only needs as an input a vector with all orig.idents.
}
