% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VDJ.utils.R
\name{gapV}
\alias{gapV}
\title{introduce IMGT gap in V sequence}
\usage{
gapV(seq, v_germ_start, v_germ_end, v_call, references)
}
\arguments{
\item{seq}{a sequence}

\item{v_germ_start}{v_germline_start output from igblast}

\item{v_germ_end}{v_germline_end output from igblast}

\item{v_call}{v_call output from igblast}

\item{references}{list of gapped reference sequence named according to possible v_calls}
}
\description{
\code{gapV} introduce IMGT gap in V sequence
}
\keyword{internal}
