% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowJo.utils.R
\name{importFJGates}
\alias{importFJGates}
\title{Imports INX FlowJo gating results and merge it to an existing AIRR formatted data frame [optional]}
\usage{
importFJGates(
  FJ_files,
  db = NULL,
  orig.ident = "sample_id",
  directory = "directory",
  filename = "filename",
  import.columns = "all",
  rm.columns = c("plate_id", "sort_id", "sort_order", "orig.ident", "donor_id",
    "time_point", "Time"),
  overwrite.columns = TRUE,
  plate_format = "96",
  Idx_to_character = list(`96` = data.frame(IdxRow = c(8, 7, 6, 5, 4, 3, 2, 1), Row =
    c("A", "B", "C", "D", "E", "F", "G", "H")), `384` = data.frame(IdxRow = c(16, 15, 14,
    13, 12, 11, 10, 9, 8, 7, 6, 5, 4, 3, 2, 1), Row = c("A", "B", "C", "D", "E", "F",
    "G", "H", "I", "J", "K", "L", "M", "N", "O", "P"))),
  export_missing_seq = FALSE
)
}
\arguments{
\item{FJ_files}{a data frame with the following minimal columns : orig.ident = "sample_id", directory = "directory", filename = "filename", for plasticity, all three colnames can be modified and entered as separate arguments}

\item{db}{an AIRR formatted data frame containing heavy and light chain sequences to which the imported FlowJo export data will be appended. if NULL, returning only the merged FlowJo export data}

\item{orig.ident}{column in FJ_files to use to find the sample orig.ident [default = "sample_id"], will be used to recreate the beginning of the cell_id (should contain all info other than the plate_id and well_id)}

\item{directory}{column in FJ_files to use to find the file directory}

\item{filename}{column in FJ_files to use to find the full name of each file to be imported}

\item{import.columns}{which columns to import from all imported files [default = "all, minus the ones removed (listed in the rm.columns parameter)]}

\item{rm.columns}{which columns to remove from all imported files}

\item{overwrite.columns}{whether to overwrite columns in the imported sanger files with data provided in the sanger_dir recap file [default = FALSE].}

\item{plate_format}{format of the plate used for sorting [default = "96"], should be one listed in Idx_to_character}

\item{Idx_to_character}{list of dataframes (one per plate format) used to convert IdxRow column in the FlowJo export file to the full well_id.}

\item{export_missing_seq}{whether to export a table with all cell_id without corresponding sequence in the provided db.}
}
\value{
if a AIRR formatted data frame is provided, the same data frame with additional columns corresponding to the exported information from the INX files processed in FlowJo.
Else, a data frame with the merged data from the imported FlowJo export files.
}
\description{
\code{importFJGates} imports a list of FlowJo export files and merge it
}
