% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VDJ.utils.R
\name{importSangerVDJ}
\alias{importSangerVDJ}
\title{Importing Sanger-based VDJ sequencing data post igblast formatting and merging it to an existing AIRR formatted data frame [optional]}
\usage{
importSangerVDJ(
  sanger_files,
  db = NULL,
  orig.ident = "sample_id",
  cell_id = "cell_id",
  sequence = "sequence",
  sequence_id = "sequence_id",
  directory = "directory",
  filename = "filename",
  rename.columns = list(list(old_name = "cell_type", new_name = "sorted_cell_type")),
  prefix.columns = c("low10QC_alternate_calls", "low30QC_alternate_calls",
    "missing_v_bp", "missing_j_bp", "comments"),
  rm.columns = c("well_id", "alternate_well_id", "plate_id", "germline_alignment_d_mask",
    "mu_count", "mu_count_cdr_r", "mu_count_cdr_s", "mu_count_fwr_r", "mu_count_fwr_s",
    "primary_sequence", "raw_sequence"),
  overwrite.columns = FALSE,
  run_igblast = FALSE,
  update_c_call = FALSE,
  seq_type = c("Ig", "TCR"),
  organism = c("human", "mouse", "rabbit", "rat", "rhesus_monkey"),
  igblast_dir = "~/share/igblast",
  imgt_dir = "~/share/germlines/imgt/",
  na.rm = TRUE
)
}
\arguments{
\item{sanger_files}{a data frame with the following minimal columns : "sample_id", "directory", "filename"}

\item{db}{an AIRR formatted data frame containing heavy and light chain sequences to which the imported sanger VDJ data will be appended. if NULL, returning only the merged sanger VDJ data}

\item{orig.ident}{column to use to set orig.ident; should be either "sample_id" (defined in sanger_dir, if importing plate by plate) or "sort_id" (defined inside each sanger file), if importing from one big table for all plates).}

\item{cell_id}{name of the column containing cell_id informations}

\item{directory}{name of the column containing cell_id informations}

\item{filename}{name of the column containing cell_id informations}

\item{rename.columns}{which columns to rename; a list of pairs new_name/old_name is expected}

\item{prefix.columns}{which columns to rename by adding teh "sanger_" prefix}

\item{rm.columns}{which columns to remove from all imported sanger files}

\item{overwrite.columns}{whether to overwrite columns in the imported sanger files with data provided in the sanger_dir recap file [default = FALSE].}

\item{run_igblast}{whether to run igblast}

\item{update_c_call}{whether to update c_call}

\item{igblast_dir}{path to igblast database}

\item{imgt_dir}{path to IMGT-gapped database}

\item{na.rm}{whether to remove wells without associated cell_id}
}
\description{
\code{importSangerVDJ} import a list of AIRR data frames, update colnames and merge with provided AIRR data frame
}
