% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VDJ.plots.R
\name{plotMultiHistogram}
\alias{plotMultiHistogram}
\title{Plot a basic histogram}
\usage{
plotMultiHistogram(df, density = TRUE, feature, label_column, vline)
}
\arguments{
\item{df}{a data frame with at least the feature column}

\item{density}{whether to add a density curve}

\item{feature}{which column to use for plotting}

\item{label_column}{column used for fill color}

\item{vline}{value for added vline}
}
\value{
a simple histogram
}
\description{
\code{plotMultiHistogram}
}
\details{
returns a simple histogram
}
