% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VDJ.utils.R
\name{reformatVDJinput}
\alias{reformatVDJinput}
\title{rename columns to harmonize with the immcantation pipeline and between technologies, also removes a few useless ones.}
\usage{
reformatVDJinput(
  db,
  tech,
  cell_id = "cell_id",
  locus = "locus",
  heavy = "IGH",
  productive = "productive",
  complete_vdj = "complete_vdj",
  sequence_id = "sequence_id",
  umi_count = "umi_count",
  consensus_count = "consensus_count",
  junction = "junction",
  junction_aa = "junction_aa",
  sequence = "sequence",
  v_call = "v_call",
  d_call = "d_call",
  j_call = "j_call",
  c_call = "c_call",
  remove_columns = list(`10X` = c("high_confidence", "raw_clonotype_id",
    "raw_consensus_id", "exact_subclonotype_id", "fwr1_aa", "fwr2_aa", "fwr3_aa",
    "fwr4_aa", "cdr1_aa", "cdr2_aa", "cdr3_aa"), BD = c("cell_type_experimental",
    "high_quality_cell", "high_quality_cell_tcr_bcr", "sequence_aa",
    "sequence_aa_length", "sequence_alignment_length", "cdr3_length", "fwr1_aa",
    "fwr2_aa", "fwr3_aa", "fwr4_aa", "cdr1_aa", "cdr2_aa", "cdr3_aa",
    "germline_alignment_aa", "v_germline_alignment", "v_germline_alignment_aa",
    "d_germline_alignment", 
     "d_germline_alignment_aa", "j_germline_alignment",
    "j_germline_alignment_aa"))
)
}
\arguments{
\item{db}{a data frame directly imported from SevenBridges (BD) or CellRanger (10X).}

\item{tech}{name of the scRNA-seq technology used (one of 10X or BD)}

\item{locus}{name of the column containing locus informations}

\item{heavy}{which vlue to use in locus to identify heavy chains [default: IGH for BCR]}

\item{productive}{name of the column containing contig productiveness informations}

\item{complete_vdj}{name of the column containing contig completness informations}

\item{sequence_id}{name of the column containing sequence_ids}

\item{umi_count}{name of the column containing umi_count informations}

\item{consensus_count}{name of the column containing consensus_count informations}

\item{junction}{name of the column containing identified junctions (nucleotide format)}

\item{junction_aa}{name of the column containing identified junctions (amino-acid format)}

\item{sequence}{name of the column containing full sequences}

\item{v_call}{name of the column containing v_call informations}

\item{d_call}{name of the column containing d_call informations}

\item{j_call}{name of the column containing j_call informations}

\item{c_call}{name of the column containing c_call informations}

\item{remove_columns}{name of collumns to remove from original 10X or BD data frames}

\item{cell_id.}{name of the column containing cell_ids}
}
\value{
a reformatted dataframe with same name for both 10X or BD technologies. Re-running igblast is however advised to get a full AIRR formatted dataframe.
}
\description{
\code{reformatVDJinput}
}
\details{
the following columns are unique to the SevenBridges pipeline from BD:
"cell_type_experimental"; "high_quality_cell"; "complete_vdj"; "dominant"; "putative_cell"
"sequence_length"; "sequence_aa"; "sequence_aa_length" ; "sequence_alignment_length"
"sequence_alignment_aa"; "sequence_alignment_aa_length"; "cdr3_length"; "fwr1_aa";"fwr2_aa"; "fwr3_aa"; "fwr4_aa"; "cdr1_aa"; "cdr2_aa"; "cdr3_aa"
"germline_alignment_aa"; "v_germline_alignment"; "v_germline_alignment_aa"; "d_germline_alignment"; "d_germline_alignment_aa"; "j_germline_alignment";"j_germline_alignment_aa"
some will be added or modified at the CreateGermline() step later after clone inference (current germline alignment is no gapped so not usable in observedMutations)
https://bd-rhapsody-bioinfo-docs.genomics.bd.com/outputs/outputs_vdj_contigs.htm

the following columns are named differently or unique to the 10x pipeline from BD:
barcode, is_cell,	contig_id,	high_confidence,	length,	chain,	v_gene,	d_gene,	j_gene,	c_gene,
full_length,	fwr1,	fwr1_nt,	cdr1,	cdr1_nt,	fwr2,	fwr2_nt,	cdr2,	cdr2_nt,	fwr3,	fwr3_nt,	cdr3,	cdr3_nt,	fwr4,	fwr4_nt,
reads,	umis,	raw_clonotype_id,	raw_consensus_id,	exact_subclonotype_id,
}
