% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VDJ.utils.R
\name{runAb1QC}
\alias{runAb1QC}
\title{unzip Eurofins-based results and run all steps from QC to final AIRR formatted datatable}
\usage{
runAb1QC(
  Ab1_folder,
  outfolder = NULL,
  outfilename = NULL,
  save = c("png", "html"),
  reticulate_py_env = NULL,
  primers = c("IgG", "IgM", "IgL", "IgK", "Mix"),
  trim_cutoff = NULL,
  nproc = NULL
)
}
\arguments{
\item{Ab1_folder}{path to a folder containing .Ab1 files}

\item{outfolder}{folder to output files to}

\item{save}{whether to save png or html plots}

\item{reticulate_py_env}{if you want to force a precise python environment to use for plotly. [should be "~/Library/r-miniconda-arm64/envs/r-reticulate/bin/python" if following installation instructions on MACOS 10.15.4]}

\item{primers}{primers used for PCR}

\item{trim_cutoff}{cutoff thresholds to use [if set to NULL, default to automatic cutoffs for each primers: "IgG" = 300, "IgM" = 275, "IgK" = 400, "IgL"= 275, "Mix" = 275, "Undefined" = 275]}

\item{nproc}{number of processors to use, if NULL, will be automatically set based on available processors.}
}
\description{
\code{runAb1QC}
}
\keyword{internal}
