% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VDJ.utils.R
\name{runAssignGenes}
\alias{runAssignGenes}
\title{calls Immcantation (ChangeO) AssignGenes.py and MakeDb.py scripts from within R and update some columns names depending on selected technology}
\usage{
runAssignGenes(
  db,
  type = c("nt", "aa"),
  seq_type = c("Ig", "TCR"),
  organism = c("human", "mouse", "rabbit", "rat", "rhesus_monkey"),
  igblast_dir = "~/share/igblast/",
  imgt_dir = "~/share/germlines/imgt/",
  sequence = "sequence",
  sequence_id = "sequence_id",
  output = FALSE,
  output_folder = "igblast_results",
  log = FALSE,
  log_file = NULL
)
}
\arguments{
\item{db}{a data frame containing at least a sequence and a sequence_id column.}

\item{igblast_dir}{path to igblast database [default = path suggested on installation: https://changeo.readthedocs.io/en/stable/examples/igblast.html]}

\item{imgt_dir}{path to imgt-gapped database [default = path suggested on installation: https://changeo.readthedocs.io/en/stable/examples/igblast.html]}

\item{sequence}{name of the column containing the original sequence.}

\item{sequence_id}{name of the column containing sequence identifier.}

\item{output}{whether to output graphs with umi_counts for dominant versus second IGH VDJ contig and the recap excel workbook. If set to FALSE, only the corrected database is returned.}

\item{output_folder}{name of the folder in which graph and recap excel workbooks will be saved [default = "igblast_results/"].}

\item{log}{whether to create a log file}

\item{log_file}{name of the log file}
}
\value{
a named list of AIRR formatted data frame corresponding to the pass and fail results of Igblast reformatted using Makedb and all additional columns from the submitted data frame.
}
\description{
\code{runAssignGenes} old pipeline to run igblast
}
\details{
1. will first run igblast
2. import igblast results and reformat data frame to return all metadata from the original dataframe
dependencies: standalone igblast; Changeo (AssignGenes.py, MakeDb.py); dplyr; readr; Biostrings
}
