% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VDJ.utils.R
\name{runBlastnC}
\alias{runBlastnC}
\title{call C regions from VDJC sequences}
\usage{
runBlastnC(
  db,
  seq_type = c("Ig", "TCR"),
  organism = c("human", "mouse", "rabbit", "rat", "rhesus_monkey"),
  igblast_dir = "~/share/igblast",
  sequence = "sequence",
  sequence_id = "sequence_id",
  c_call = "c_call",
  output = FALSE,
  output_folder = "C_results",
  nproc_blast = 10
)
}
\arguments{
\item{db}{a AIRR formated data frame containing at least a sequence and a sequence_id column.}

\item{seq_type}{type of VDJ sequence ("Ig" or "TCR" to match igblastb requirements)}

\item{organism}{organism (any of "human", "mouse", "rhesus_monkey; for other see https://changeo.readthedocs.io/en/stable/examples/igblast.html)}

\item{igblast_dir}{path to igblast database [default = path suggested on installation: https://changeo.readthedocs.io/en/stable/examples/igblast.html]}

\item{sequence}{name of the column containing the original sequence.}

\item{sequence_id}{name of the column containing sequence identifier.}

\item{c_call}{name of the column containing c_call.}

\item{output}{whether to output graphs with umi_counts for dominant versus second IGH VDJ contig and the recap excel workbook. If set to FALSE, only the corrected database is returned.}

\item{output_folder}{name of the folder in which graph and recap excel workbooks will be saved [default = "VDJ_QC"].}
}
\value{
the inputted dataframe with an additional c_call column after the j_call column.
}
\description{
\code{runBlastnC} run Blastn on a Blastable IMGT C region datbase return a corrected c_call
}
\details{
1. will first run Blastn
2. import Blastn results and select the alignment(s) with the highest scores
3. in case of multiple gene calls with identical score, will return an agregated call (e.g. (IGHG1|IGHG2))
}
