% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VDJ.utils.R
\name{runIgblastn}
\alias{runIgblastn}
\title{directly calls igblastn from within R and outputs an airr formatted table}
\usage{
runIgblastn(
  db,
  sequence = "sequence",
  sequence_id = "sequence_id",
  seq_type = c("Ig", "TCR"),
  organism = c("human", "mouse", "rabbit", "rat", "rhesus_monkey"),
  igblast_dir = "~/share/igblast/",
  imgt_dir = "~/share/germlines/imgt/",
  gapV = TRUE,
  output = FALSE,
  output_folder = "igblast_results",
  log = FALSE,
  log_file = NULL,
  nproc_igblast = 10
)
}
\arguments{
\item{db}{a data frame containing at least a sequence and a sequence_id column.}

\item{sequence}{name of the column containing the original sequence.}

\item{sequence_id}{name of the column containing sequence identifier.}

\item{seq_type}{type of VDJ sequence ("Ig" or "TCR" to match igblastb requirements)}

\item{organism}{organism (any of "human", "mouse", "rhesus_monkey; for other see https://changeo.readthedocs.io/en/stable/examples/igblast.html)}

\item{igblast_dir}{path to igblast database [default = path suggested on installation: https://changeo.readthedocs.io/en/stable/examples/igblast.html]}

\item{imgt_dir}{path to imgt-gapped database [default = path suggested on installation: https://changeo.readthedocs.io/en/stable/examples/igblast.html]}

\item{output}{whether to output graphs with umi_counts for dominant versus second IGH VDJ contig and the recap excel workbook. If set to FALSE, only the corrected database is returned.}

\item{output_folder}{name of the folder in which graph and recap excel workbooks will be saved [default = "igblast_results/"].}

\item{log}{whether to create a log file}

\item{log_file}{name for the log file}

\item{nproc_igblast}{number of proc to use for igblast}
}
\value{
a named list of AIRR formatted data frame corresponding to the pass and fail results of Igblast reformatted using Makedb and all additional columns from the submitted data frame.
}
\description{
\code{runIgblastn} run igblast
}
\details{
1. will first run igblast and request an airr format output
2. import igblast results and add all metadata from the original dataframe
dependencies: standalone igblastn; dplyr; readr; Biostrings; parallel,
}
