% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VDJ.utils.R
\name{vdjQCplot}
\alias{vdjQCplot}
\title{plot QC graph}
\usage{
vdjQCplot(
  db,
  use_chain = "all",
  split.by = NULL,
  type = c("vdj_doublet", "nonB_vdj_doublet"),
  output_folder = NULL,
  analysis_name = "All_sequences",
  variable_cutoff = TRUE,
  low_cutfoff = 10,
  high_cutoff = 250,
  plot_cutoff = FALSE,
  na_name = "unknown",
  colour.by = list(vdj_doublet = "is.VDJ_doublet.confidence", nonB_vdj_doublet =
    "is.nonB_VDJ_doublet.confidence"),
  colour_code = list(vdj_doublet = c(high = "darkred", low = "darkorange",
    `not a doublet` = "darkgreen"), nonB_vdj_doublet = c(high = "darkred", low =
    "darkorange", `ambient RNA` = "darkgreen", `B cell` = "grey")),
  locus = "locus",
  heavy = "IGH",
  cell_id = "cell_id",
  umi_count = "umi_count",
  second_umi_count = "second_umi_count",
  export = c("graph", "pdf")
)
}
\arguments{
\item{db}{a data table}

\item{use_chain}{which chain(s) to use}

\item{split.by}{name of the column to use to group cells and learn dataset-specific distributions}

\item{type}{whether to plot graphs related to B/B doublets or B/non-B doublets}

\item{output_folder}{name of the folder in which graph and recap excel workbooks will be saved [default = "VDJ_QC"].}

\item{analysis_name}{suffix for final plot(s) name}

\item{variable_cutoff}{whether to use dataset-specific cut-offs}

\item{low_cutfoff}{if fixed cutoffs, value to use to mark low-probability doublets}

\item{high_cutoff}{if fixed cutoffs, value to use to mark high probability heavy chain doublets.}

\item{plot_cutoff}{whether to plot cut_off on graph}

\item{na_name}{name to use for NA values in colour.by}

\item{colour.by}{parameter to use for coloring of dots}

\item{colour_code}{colors to use for coloring of dots}

\item{locus}{name of the column containing locus informations}

\item{heavy}{locus value to use for heavy chain identification}

\item{cell_id}{name of the column containing cell_ids}

\item{umi_count}{name of the column containing umi_count informations}

\item{second_umi_count}{name of the column containing second_umi_count informations}

\item{export}{what to export: graph (return the ggplot2 object) or pdf (save)}
}
\value{
a pdf plot in the output folder if export = "pdf" and/or a ggplot object if export = "graph"
}
\description{
\code{vdjQCplot} plot QC graph
}
\keyword{internal}
