% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VDJ.utils.R
\name{binContigs}
\alias{binContigs}
\title{Resolve VDJ chains multiplets, eventually using prior clustering knowledge}
\usage{
binContigs(
  data,
  plot = TRUE,
  title = "Histogram + Density Plot with Antimode-Based Bins",
  file = "DensityPlot.pdf"
)
}
\arguments{
\item{data}{vector of values.}

\item{plot}{whether to export plot}

\item{title}{title of plot}

\item{file}{filename for plot}
}
\value{
a list containing three slots:
"bins": the allocated bin for each initial value
"plot": the histogram + density plot with antimode-based bins highlighted (Kernel Density Estimation)
"top75%": the top bins corresponding to >75% of all values.
}
\description{
\code{resolveMultiContigs} resolve cases of multiple heavy or light chains within a cell
}
\keyword{internal}
