% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VDJ.plots.R
\name{DonutPlotClonotypes}
\alias{DonutPlotClonotypes}
\title{Plots multiple donut plots and save them as pdf}
\usage{
DonutPlotClonotypes(
  db,
  split.by = NULL,
  prefix = NULL,
  return_plot = FALSE,
  ...
)
}
\arguments{
\item{db}{an AIRR formatted dataframe containing bcr (heavy and light chains) or tcr (TCRA, TCRB, TCRG or TCRD) sequences. Should contain only one chain for each type per cell_id, if not run resolveMultiHC() first.}

\item{split.by}{name of column to use to group sequence when calculating clone size and frequencies.}

\item{prefix}{prefix to use for saved files}

\item{...}{arguments to be passed to SingleDonutPlotClonotypes()}
}
\value{
saved pdf for as many donut plots as there are groups in the dataframe as defined by the split.by argument.
if the split.by argument points to more than one column, will group the dataset based on the first element(s) of the split.by argument
and keep only the last as split.by argument of the SingleDonutPlotClonotypes function.
}
\description{
\code{DonutPlotClonotypes} Plots multiple donut plots and save them as pdf
}
\details{
db should be an AIRR formated database with a defined "clone_id" column
Defines clones to plot and colors based on clones properties:
Expanded clone = clone size >1 in one donor
Shared clone = clone found in at least two donors
Persisting clone = clone found in at least two time points and/or pop
Color choice:
Single Greys scale based on clone size in that donor (to account for sampling): "white" = singlets, "black"= biggest clone in all donors
Color for shared clones between donors
highlight = c("expanded", "top5")
Consistent color throughout time-points for persisting-expanded clones
Grey for non-persisting expanded clones
White for non expanded clones and regroup all sequences found only once in the dataset (non-persisting/non-expanded)
}
