% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VDJ.utils.R
\name{flagVDJdoublets}
\alias{flagVDJdoublets}
\title{calls successively flagBdoublets and flagNonBdoublets}
\usage{
flagVDJdoublets(
  db,
  analysis_name = "All_sequences",
  homotypic = TRUE,
  heterotypic = TRUE,
  output = TRUE,
  output_folder = "VDJ_QC",
  cell_id = "cell_id",
  locus = "locus",
  heavy = "IGH",
  umi_count = "umi_count",
  use_chain = c("all", "IGH"),
  split.by = NULL,
  variable_cutoff = TRUE,
  low_cutoff = 10,
  high_cutoff = 250,
  azimuth.ref = c("pbmcref", "tonsilref", "bonemarrowref"),
  azimuth.column = list(pbmcref = "predicted.celltype.l2", tonsilref =
    "predicted.celltype.l1", bonemarrowref = "predicted.celltype.l2"),
  azimuth.Bcelltypes = list(pbmcref = c("B intermediate", "B memory", "B naive",
    "Plasmablast"), tonsilref = c("B activated", "B memory", "B naive", "preGCB", "PB",
    "PC", "PC/doublet", "preMBC/doublet", "prePB", "Cycling DZ GCB", "DZ GCB",
    "DZtoLZ GCB transition", "FCRL4/5+ B memory", "LZ GCB", "LZtoDZ GCB transition"),
    bonemarrowref = c("Memory B", "Naive B", "Plasma", "pro B", "pre B",
    "transitional B")),
  scRNAseq.tech = c("10X", "BD"),
  assay = "assay",
  productive = "productive",
  complete_vdj = "complete_vdj",
  sequence_id = "sequence_id",
  consensus_count = "consensus_count",
  junction = "junction",
  junction_aa = "junction_aa",
  sequence = "sequence",
  v_call = "v_call",
  d_call = "d_call",
  j_call = "j_call",
  c_call = "c_call"
)
}
\arguments{
\item{db}{a AIRR formatted dataframe containing heavy and light chain sequences, cell_ids and umi_counts.}

\item{analysis_name}{name to use for outputs prefixes.}

\item{homotypic}{whether to run flagBdoublets()}

\item{heterotypic}{whether to run flagNonBdoublets()}

\item{output}{whether to output graphs with umi_counts for dominant versus second IGH VDJ contig and the recap excel workbook. Can be set to none, partial (only pdf + log file with QC parameters) or total (pdf + log file with QC parameters + tsv file with sequences).}

\item{output_folder}{name of the folder in which graph and recap excel workbooks will be saved [default = "VDJ_QC"].}

\item{cell_id}{name of the column containing cell identifier.}

\item{locus}{name of column containing locus values.}

\item{heavy}{which locus value to use as heavy chain [default: IGH]}

\item{umi_count}{name of the column containing the number of unique molecules (UMI) for this contig. Previously called "duplicate_count" in an earlier AIRR standard}

\item{use_chain}{which chain to use to flag VDJ doublets, can de set to "IGH" or "all" [default = "all"].}

\item{split.by}{which column to use to group cells when learning dataset-specific distributions}

\item{variable_cutoff}{whether to use variable cutoffs between datasets}

\item{low_cutoff}{if fixed cutoffs, cutoff to use to define low-probability doublets}

\item{high_cutoff}{if fixed cutoffs, cutoff to use to define high-probability doublets}

\item{azimuth.ref}{reference used when running azimuth}

\item{azimuth.column}{name of the column where azimuth calls are stored}

\item{azimuth.Bcelltypes}{name of the B cell populations in the used azimuth reference (default for pbmc.ref, tonsil.ref and bonemarrow.ref are provided)}

\item{scRNAseq.tech}{scRNA-seq technology used (one of 10X or BD)}

\item{assay}{name of the column containing assay information.}

\item{productive}{name of the column containing sequence productivity information.}

\item{complete_vdj}{name of the column containing sequence completeness information.}

\item{sequence_id}{name of the column containing the sequence_id.}

\item{consensus_count}{name of the column containing consensus count.}

\item{junction}{name of the column containing junction (nucleotide format).}

\item{junction_aa}{name of the column containing junction (amino-acid format).}

\item{sequence}{name of the column containing the full sequence.}

\item{v_call}{name of the column containing the v_call.}

\item{d_call}{name of the column containing the d_call.}

\item{j_call}{name of the column containing the j_call.}

\item{c_call}{name of the column containing the c_call.}
}
\description{
\code{flagVDJdoublets} flag potential B and nonB doublets.
}
