% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VDJ.utils.R
\name{flagVDJdoublets}
\alias{flagVDJdoublets}
\title{calls successively flagBdoublets and flagNonBdoublets}
\usage{
flagVDJdoublets(
  db,
  analysis_name = "All_sequences",
  homotypic = TRUE,
  heterotypic = TRUE,
  seq_type = c("Ig", "TCR"),
  heavy = NULL,
  light = NULL,
  output = TRUE,
  output_folder = "VDJ_QC",
  split.by = NULL,
  ref = c("azimuth.pbmcref", "azimuth.tonsilref", "azimuth.bonemarrowref"),
  ref.column = list(azimuth.pbmcref = "predicted.celltype.l2", azimuth.tonsilref =
    "predicted.cell_type.l1", azimuth.bonemarrowref = "predicted.celltype.l2"),
  ref.Bcelltypes = list(azimuth.pbmcref = c("B intermediate", "B memory", "B naive",
    "Plasmablast"), azimuth.tonsilref = c("B activated", "B memory", "B naive", "preGCB",
    "PB", "PC", "PC/doublet", "preMBC/doublet", "prePB", "Cycling DZ GCB", "DZ GCB",
    "DZtoLZ GCB transition", "FCRL4/5+ B memory", "LZ GCB", "LZtoDZ GCB transition"),
    azimuth.bonemarrowref = c("Memory B", "Naive B", "Plasma", "pro B", "pre B",
    "transitional B")),
  ref.Tcelltypes = list(azimuth.pbmcref = c("CD4 CTL", "CD4 Naive", "CD4 Proliferating",
    "CD4 TCM", "CD4 TEM", "Treg", "CD8 Naive", "CD8 Proliferating", "CD8 TCM", "CD8 TEM",
    "dnT", "gdT", "MAIT"), azimuth.tonsilref = c("CD4 naive", "CD4 Non-TFH", "CD4 TCM",
    "CD4 TFH", "CD4 TFH Mem", "CD4 TREG", "CD8 naive", "CD8 T", "CD8 TCM", "Cycling T",
    "dnT", "MAIT/TRDV2+ gdT", "non-TRDV2+ gdT"), azimuth.bonemarrowref =
    c("CD4 Effector", "CD4 Memory", "CD4 Naive", "CD8 Effector 1", "CD8 Effector 2",
    "CD8 Effector 3", "CD8 Memory", 
     "CD8 Naive", "MAIT", "T proliferating")),
  save_plot = c("pdf", "png"),
  save_tsv = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{db}{a AIRR formatted dataframe containing heavy and light chain sequences, cell_ids and umi_counts.}

\item{analysis_name}{name to use for outputs prefixes.}

\item{homotypic}{whether to run homotypicVDJdoublets()}

\item{heterotypic}{whether to run heterotypicVDJdoublets()}

\item{seq_type}{type of VDJ sequence ("Ig" or "TCR" to match igblastb requirements)}

\item{heavy}{heavy chains to be used. if set to NULL, will default to IGH for Ig and TRB/TRD for TCR.}

\item{light}{light chains to be used. if set to NULL, will default to IGL/IGK for Ig and TRA/TRG for TCR.}

\item{output}{whether to output graphs with umi_counts for dominant versus second IGH VDJ contig and the recap excel workbook. Can be set to none, partial (only pdf + log file with QC parameters) or total (pdf + log file with QC parameters + tsv file with sequences).}

\item{output_folder}{name of the folder in which graph and recap excel workbooks will be saved [default = "VDJ_QC"].}

\item{split.by}{which column to use to group cells when learning dataset-specific distributions}

\item{save_plot}{format to save the final ggplot2 object. "pdf" or "png". if set to any other value, no plot will be saved.}

\item{verbose}{whether to print recap informations to the console}

\item{...}{arguments to pass to homotypicVDJdoublets or heterotypicVDJdoublets.}
}
\description{
\code{flagVDJdoublets} flag potential B and nonB doublets.
}
