% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VDJ.plots.R
\name{HexmapClonotypes}
\alias{HexmapClonotypes}
\title{Plots individual hexmap plots and save as pdf}
\usage{
HexmapClonotypes(
  data,
  ordered = TRUE,
  radius = 1,
  padding = 0.2,
  fill_col = "origin",
  cell_id = "cell_id",
  clone_id = "clone_id",
  title = "All_sequences",
  shape = c("hex", "circle"),
  palette = NULL,
  seed = 42,
  max_colors = 10,
  return_coords = FALSE
)
}
\arguments{
\item{ordered}{whether to order clones based on size and highlighing parameter}

\item{radius}{shape radius, passed to packcircles}

\item{padding}{padding, passed to packcircles}

\item{cell_id}{name of the column containing cell identifier.}

\item{clone_id}{name of the column containing cell identifier.}

\item{shape}{shape to use [default: hex]}

\item{seed}{}

\item{db}{an AIRR formatted dataframe containing bcr (heavy and light chains) or tcr (TCRA, TCRB, TCRG or TCRD) sequences. Should contain only one chain for each type per cell_id, if not run resolveMultiHC() first.}

\item{split.by}{name of column to use to group sequence when calculating clone size and frequencies.}

\item{highlight}{name of the column use to color each clone [default: "c_call].}

\item{highlight_col}{color scheme to use for coloring clones}

\item{prefix}{prefix to use for saved files}

\item{use_chain}{which chain to use [default: "IGH"], each cell should only have one contig for this chain}

\item{locus}{name of column containing locus values.}

\item{groups_to_plot}{which groups to plot}

\item{col.line}{color for lines in donut plot}

\item{plots_folder}{name for export folder [default: "Donut_plots"]}

\item{productive}{name of column containing productive calls.}

\item{productive_only}{whether to exclude non productive sequences [default: TRUE]}
}
\value{
a hexmap plot
}
\description{
\code{HexmapClonotypes} Plots individual hexmap plots and save as pdf
}
