% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VDJ.utils.R
\name{runAb1QC}
\alias{runAb1QC}
\title{unzip Eurofins-based results and run all steps from QC to final AIRR formatted datatable}
\usage{
runAb1QC(
  Ab1_folder,
  outfolder = NULL,
  outfilename = NULL,
  save = c("png", "html"),
  primers = c("IgG", "IgM", "IgL", "IgK", "Mix"),
  trim_cutoff = list(IgG = 300, IgM = 275, IgK = 400, IgL = 275, Mix = 275, Undefined =
    275),
  nproc = NULL
)
}
\arguments{
\item{Ab1_folder}{path to a folder containing .Ab1 files}

\item{outfolder}{folder to output files to}

\item{save}{whether to save png or html plots}

\item{primers}{primers used for PCR}

\item{trim_cutoff}{cutoff thresholds to use depending on primers}

\item{nproc}{number of processors to use, if NULL, will be automatically set based on available processors.}
}
\description{
\code{runAb1QC}
}
\keyword{internal}
