% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VDJ.utils.R
\name{scFindBCRClones}
\alias{scFindBCRClones}
\title{Full BCR analysis pipeline for heavy chain curated AIRR single-cell repertoire datasets}
\usage{
scFindBCRClones(db, ...)
}
\arguments{
\item{db}{an AIRR formatted dataframe containing heavy and light chain sequences. Should contain only one heavy chain (IGH) per cell_id, if not run resolveMultiHC() first.}

\item{...}{arguments to be passed to scFindClones}
}
\value{
an AIRR formatted dataframe with a new clone_id column and clone attribution for all BCR related data (see scFindClones).
}
\description{
\code{scFindBCRlones} performs clonal clustering of single cell data
}
\details{
wrapper to run scFindClones with BCR specific arguments. All other arguments from scFindClones can also be used.
}
